+++ apps/webapp/app/routes/admin.tsx
@@ -48,6 +48,10 @@ export default function Page() {
               to: "/admin/back-office",
               end: false,
             },
+            {
+              label: "Data Stores",
+              to: "/admin/data-stores",
+            },
           ]}
           layoutId={"admin"}
         />